+++ docs/changelog.rst
@@ -129,7 +129,7 @@ Change Log
 
 0.2.2
 -----
-* Fixed the _resolve_url to not assume running on **nix system.
+* Fixed the _resolve_url to not assume running on **nix** system.
 
 0.2.1
 -----